analysis/requirements.txt
awscli==1.11.141
backports-abc==0.4
backports.functools-lru-cache==1.3
backports.ssl-match-hostname==3.5.0.1
basemap==1.0.7
botocore==1.6.8
Cartopy==0.14.2
cassandra-driver==3.7.1
certifi==2016.2.28
colorama==0.3.7
configobj==5.0.6
cycler==0.10.0
Cython==0.24
descartes==1.0.2
docutils==0.14
funcsigs==1.0.2
futures==3.1.1
GDAL==2.0.0
Jinja2==2.8
jmespath==0.9.3
MarkupSafe==0.23
matplotlib==1.5.1
mock==2.0.0
mpld3==0.2
netCDF4==1.2.4
nexus-data-access==0.32
nexusproto==0.3
numpy==1.11.2
pbr==1.10.0
protobuf==2.6.1
pyasn1==0.3.3
pyface==4.5.2
Pygments==2.1.3
pyparsing==2.1.4
pyproj==1.9.5.1
pyshp==1.2.10
python-dateutil==2.6.1
pytz==2016.6.1
PyYAML==3.12
requests==2.11.0
rsa==3.4.2
s3transfer==0.1.10
scipy==0.17.1
Shapely==1.5.16
singledispatch==3.4.0.3
six==1.10.0
solrpy==0.9.7
tornado==4.3
traits==4.5.0
traitsui==4.5.1
utm==0.4.1
